Refactored drop zone
